requirements/test.txt
-r base.txt

django-nose
